Move single state utils back into single state tests file
